Fix vuepress-v2.0.0-beta49
- change verson to v2.0.0-beta49
- detele clientAppSetup.js and merge into client.js
- modify clientConfigFile
- modify defineClientConfig
