+++ packages/@snippetors/vuepress-plugin-code-copy/lib/client.js
@@ -1,7 +1,71 @@
+import { defineClientConfig } from "@vuepress/client";
+import { createApp, onUpdated, onMounted, onBeforeUnmount, watch } from "vue";
+import { usePageData } from "@vuepress/client";
+
 import CodeCopy from "./components/CodeCopy.vue";
 
-import { defineClientAppEnhance } from "@vuepress/client";
+import "./theme/style.css";
 
-export default defineClientAppEnhance(({ app, router, siteData }) => {
-  app.component("CodeCopy", CodeCopy);
+export default defineClientConfig({
+  enhance({ app }) {
+    app.component("CodeCopy", CodeCopy);
+  },
+  setup() {
+    console.log(snippetorsCodeCopyOptions);
+    const page = usePageData();
+
+    const update = () => {
+      setTimeout(() => {
+        document
+          .querySelectorAll(snippetorsCodeCopyOptions.selector)
+          .forEach((el) => {
+            if (el.classList.contains("code-copy-added")) return;
+            let options = {
+              align: snippetorsCodeCopyOptions.align,
+              color: snippetorsCodeCopyOptions.color,
+              backgroundTransition:
+                snippetorsCodeCopyOptions.backgroundTransition,
+              backgroundColor: snippetorsCodeCopyOptions.backgroundColor,
+              successText: snippetorsCodeCopyOptions.successText,
+              successTextColor: snippetorsCodeCopyOptions.successTextColor,
+              staticIcon: snippetorsCodeCopyOptions.staticIcon,
+            };
+            let instance = createApp(CodeCopy, {
+              parent: el,
+              code: el.querySelector("pre").innerText,
+              options: options,
+            });
+            let childEl = document.createElement("div");
+            el.appendChild(childEl);
+            instance.mount(childEl);
+  
+            el.classList.add("code-copy-added");
+          });
+      }, 500); // should be greater than the page transition duration https://github.com/vuepress/vuepress-next/blob/13ab3906728a8356402c91d8ecdf61152b842d8e/packages/%40vuepress/theme-default/src/client/styles/transitions.scss#L12-L17
+    };
+  
+    onMounted(() => {
+      console.log("is mounted")
+      update();
+      window.addEventListener(
+        "snippetors-vuepress-plugin-code-copy-update-event",
+        update
+      );
+    });
+  
+    onBeforeUnmount(() => {
+      console.log("UnMounted")
+      window.removeEventListener(
+        "snippetors-vuepress-plugin-code-copy-update-event",
+        update
+      );
+    });
+  
+    onUpdated(() => {
+      console.log("UPDATE!!")
+      update();
+    });
+  
+    watch(() => page.value.path, update);
+  },
 });
