+++ packages/@snippetors/vuepress-plugin-tabs/lib/index.js
@@ -3,7 +3,7 @@ const { path } = require("@vuepress/utils");
 module.exports = (options, app) => ({
   name: "@snippetors/vuepress-plugin-tabs",
   multiple: false,
-  clientAppEnhanceFiles: [path.resolve(__dirname, "./client.js")],
+  clientConfigFile: path.resolve(__dirname, "./client.js"),
   bundlerConfig: {
     scss: {
       includePath: ["./theme/tabs.scss"],
